Use relative cache path; guard rimraf

+++ Makefile
@@ -1,8 +1,7 @@
-CACHE_DIR = $(abspath .cache)
-CACHE_BIN := $(CACHE_DIR)/bin
+CACHE_DIR = .cache
 SHELL := /usr/bin/env bash -o pipefail
 .DEFAULT_GOAL = all
-export PATH := $(CACHE_BIN):$(PATH)
+export PATH := $(abspath $(CACHE_DIR)/bin):$(PATH)
 
 # TODO remove after the repository has been made public
 export GOPRIVATE := github.com/bufbuild/protobuf-es
@@ -80,21 +79,21 @@ LICENSE_HEADER_YEAR_RANGE := 2021-2022
 LICENSE_HEADER_IGNORES := .cache\/ node_module\/ packages\/bench-codesize\/src\/gen\/ packages\/connect-web\/dist\/
 LICENSE_HEADER_DEP := $(CACHE_DIR)/dep/license-header-$(LICENSE_HEADER_VERSION)
 $(LICENSE_HEADER_DEP):
-	GOBIN=$(CACHE_BIN) go install github.com/bufbuild/buf/private/pkg/licenseheader/cmd/license-header@$(LICENSE_HEADER_VERSION)
+	GOBIN=$(abspath $(CACHE_DIR)/bin) go install github.com/bufbuild/buf/private/pkg/licenseheader/cmd/license-header@$(LICENSE_HEADER_VERSION)
 	mkdir -p $(dir $(LICENSE_HEADER_DEP)) && touch $(LICENSE_HEADER_DEP)
 
 # Install git-ls-files-unstaged
 GIT_LS_FILES_UNSTAGED_VERSION ?= v1.1.0
 GIT_LS_FILES_UNSTAGED_DEP := $(CACHE_DIR)/dep/git-ls-files-unstaged-$(GIT_LS_FILES_UNSTAGED_VERSION)
 $(GIT_LS_FILES_UNSTAGED_DEP):
-	GOBIN=$(CACHE_BIN) go install github.com/bufbuild/buf/private/pkg/git/cmd/git-ls-files-unstaged@$(GIT_LS_FILES_UNSTAGED_VERSION)
+	GOBIN=$(abspath $(CACHE_DIR)/bin) go install github.com/bufbuild/buf/private/pkg/git/cmd/git-ls-files-unstaged@$(GIT_LS_FILES_UNSTAGED_VERSION)
 	mkdir -p $(dir $(GIT_LS_FILES_UNSTAGED_DEP)) && touch $(GIT_LS_FILES_UNSTAGED_DEP)
 
 # Install golangci-lint
 GOLANGCI_LINT_VERSION ?= v1.44.0
 GOLANGCI_LINT_DEP := $(CACHE_DIR)/dep/golangci-lint-$(GOLANGCI_LINT_VERSION)
 $(GOLANGCI_LINT_DEP):
-	GOBIN=$(CACHE_BIN) go install github.com/golangci/golangci-lint/cmd/golangci-lint@$(GOLANGCI_LINT_VERSION)
+	GOBIN=$(abspath $(CACHE_DIR)/bin) go install github.com/golangci/golangci-lint/cmd/golangci-lint@$(GOLANGCI_LINT_VERSION)
 	mkdir -p $(dir $(GOLANGCI_LINT_DEP)) && touch $(GOLANGCI_LINT_DEP)
 
 
@@ -109,7 +108,7 @@ all: build test format lint bench-codesize ## build, test, format, lint, and ben
 clean: ## Delete build artifacts and installed dependencies
 	cd $(BENCHCODESIZE_DIR); npm run clean
 	cd $(WEB_DIR); npm run clean
-	rm -rf $(CACHE_DIR)/*
+	[ -n "$(CACHE_DIR)" ] && rm -rf $(CACHE_DIR)/*
 	rm -rf node_modules
 	rm -rf packages/protoc-gen-*/bin/*
 
@@ -136,7 +135,6 @@ bench-codesize: $(BENCHCODESIZE_GEN) node_modules $(WEB_BUILD) ## Benchmark code
 	cd $(BENCHCODESIZE_DIR) && npm run report
 
 set-version: ## Set a new version in for the project, i.e. make set-version SET_VERSION=1.2.3
-	node make/scripts/update-go-version-file.js cmd/protoc-gen-es/version.go $(SET_VERSION)
 	node make/scripts/update-go-version-file.js cmd/protoc-gen-connect-web/version.go $(SET_VERSION)
 	node make/scripts/set-workspace-version.js $(SET_VERSION)
